backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
google-generativeai==0.8.3
python-docx==1.1.2
python-pptx==1.0.2
python-dotenv==1.0.1
